Enable use of breakdown facet even if topFacets is empty
In order to use cross-team queries, a user that is a member of all
teams that will be queried must be used. In testing, this was a standard
user account that wasn't ingesting data. Since users can enter their own
value for breakdown, it shouldn't matter whether the form is pre-filled
or not.

DSET-3658